# docs/openapi.yaml
openapi: 3.1.0
info:
  title: September Sublet Tracker API
  version: 1.0.0
  description: Scoped collaborative project API. See agent-api.md for sync, leases, bulk semantics, and Hermes examples.
servers:
  - url: https://{host}/api/v1
    variables:
      host: {default: homing.hartphoenix.com}
security:
  - bearerAuth: []
paths:
...
  /me/projects:
    get:
      tags: [me, projects]
      security: [{bearerAuth: ['projects:read']}]
      responses: {'200': {description: Every project available to the user, content: {application/json: {schema: {$ref: '#/components/schemas/ProjectList'}}}}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  schemas:
...
    ProjectCreate:
      type: object
      required: [name, prompt]
      properties:
        name: {type: string, maxLength: 200}
        description: {type: string, maxLength: 5000}
        prompt: {type: string, maxLength: 20000}
        criteria: {type: object, additionalProperties: true}
...
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectCreate'
        - type: object
          required: [id, role, prompt_revision, latest_change_sequence]
          properties:
            id: {type: string, format: uuid}
            role: {type: string, enum: [owner, editor, viewer]}
            prompt_revision: {type: integer}
            latest_change_sequence: {type: integer}
            created_at: {type: string, format: date-time}
    ProjectList:
      type: object
      properties: {items: {type: array, items: {$ref: '#/components/schemas/Project'}}}
...